third_party/WebKit/LayoutTests/http/tests/devtools/console/console-viewport-control-expected.txt

CONSOLE MESSAGE: line 9: Message #0
CONSOLE MESSAGE: line 9: Message #1
CONSOLE MESSAGE: line 9: Message #2
CONSOLE MESSAGE: line 9: Message #0
CONSOLE MESSAGE: line 9: Message #1
CONSOLE MESSAGE: line 9: Message #2
CONSOLE MESSAGE: line 9: Message #3
CONSOLE MESSAGE: line 9: Message #4
CONSOLE MESSAGE: line 9: Message #5
CONSOLE MESSAGE: line 9: Message #6
CONSOLE MESSAGE: line 9: Message #7
CONSOLE MESSAGE: line 9: Message #8
CONSOLE MESSAGE: line 9: Message #9
CONSOLE MESSAGE: line 9: Message #10
CONSOLE MESSAGE: line 9: Message #11
CONSOLE MESSAGE: line 9: Message #12
CONSOLE MESSAGE: line 9: Message #0
CONSOLE MESSAGE: line 9: Message #1
CONSOLE MESSAGE: line 9: Message #2
CONSOLE MESSAGE: line 9: Message #3
CONSOLE MESSAGE: line 9: Message #4
CONSOLE MESSAGE: line 9: Message #5
CONSOLE MESSAGE: line 9: Message #6
CONSOLE MESSAGE: line 9: Message #7
CONSOLE MESSAGE: line 9: Message #8
CONSOLE MESSAGE: line 9: Message #9
CONSOLE MESSAGE: line 9: Message #10
CONSOLE MESSAGE: line 9: Message #11
CONSOLE MESSAGE: line 9: Message #12
CONSOLE MESSAGE: line 9: Message #13
CONSOLE MESSAGE: line 9: Message #14
CONSOLE MESSAGE: line 9: Message #15
CONSOLE MESSAGE: line 9: Message #16
CONSOLE MESSAGE: line 9: Message #17
CONSOLE MESSAGE: line 9: Message #18
CONSOLE MESSAGE: line 9: Message #19
CONSOLE MESSAGE: line 9: Message #20
CONSOLE MESSAGE: line 9: Message #21
CONSOLE MESSAGE: line 9: Message #22
CONSOLE MESSAGE: line 9: Message #23
CONSOLE MESSAGE: line 9: Message #24
CONSOLE MESSAGE: line 9: Message #0
CONSOLE MESSAGE: line 9: Message #1
CONSOLE MESSAGE: line 9: Message #2
CONSOLE MESSAGE: line 9: Message #3
CONSOLE MESSAGE: line 9: Message #4
CONSOLE MESSAGE: line 9: Message #5
CONSOLE MESSAGE: line 9: Message #6
CONSOLE MESSAGE: line 9: Message #7
CONSOLE MESSAGE: line 9: Message #8
CONSOLE MESSAGE: line 9: Message #9
CONSOLE MESSAGE: line 9: Message #10
CONSOLE MESSAGE: line 9: Message #11
CONSOLE MESSAGE: line 9: Message #12
CONSOLE MESSAGE: line 9: Message #13
CONSOLE MESSAGE: line 9: Message #14
CONSOLE MESSAGE: line 9: Message #15
CONSOLE MESSAGE: line 9: Message #16
CONSOLE MESSAGE: line 9: Message #17
CONSOLE MESSAGE: line 9: Message #18
CONSOLE MESSAGE: line 9: Message #19
CONSOLE MESSAGE: line 9: Message #20
CONSOLE MESSAGE: line 9: Message #21
CONSOLE MESSAGE: line 9: Message #22
CONSOLE MESSAGE: line 9: Message #23
CONSOLE MESSAGE: line 9: Message #24
CONSOLE MESSAGE: line 9: Message #0
CONSOLE MESSAGE: line 9: Message #1
CONSOLE MESSAGE: line 9: Message #2
CONSOLE MESSAGE: line 9: Message #0
CONSOLE MESSAGE: line 9: Message #1
CONSOLE MESSAGE: line 9: Message #2
CONSOLE MESSAGE: line 9: Message #3
CONSOLE MESSAGE: line 9: Message #4
CONSOLE MESSAGE: line 9: Message #5
CONSOLE MESSAGE: line 9: Message #6
CONSOLE MESSAGE: line 9: Message #7
CONSOLE MESSAGE: line 9: Message #8
CONSOLE MESSAGE: line 9: Message #9
CONSOLE MESSAGE: line 9: Message #10
CONSOLE MESSAGE: line 9: Message #11
CONSOLE MESSAGE: line 9: Message #12
CONSOLE MESSAGE: line 9: Message #13
CONSOLE MESSAGE: line 9: Message #14
CONSOLE MESSAGE: line 9: Message #15
CONSOLE MESSAGE: line 9: Message #16
CONSOLE MESSAGE: line 9: Message #17
CONSOLE MESSAGE: line 9: Message #18
CONSOLE MESSAGE: line 9: Message #19
CONSOLE MESSAGE: line 9: Message #20
CONSOLE MESSAGE: line 9: Message #21
CONSOLE MESSAGE: line 9: Message #22
CONSOLE MESSAGE: line 9: Message #23
CONSOLE MESSAGE: line 9: Message #24
CONSOLE MESSAGE: line 9: Message #0
CONSOLE MESSAGE: line 9: Message #1
CONSOLE MESSAGE: line 9: Message #2
CONSOLE MESSAGE: line 9: Message #3
CONSOLE MESSAGE: line 9: Message #4
CONSOLE MESSAGE: line 9: Message #5
CONSOLE MESSAGE: line 9: Message #6
CONSOLE MESSAGE: line 9: Message #7
CONSOLE MESSAGE: line 9: Message #8
CONSOLE MESSAGE: line 9: Message #9
CONSOLE MESSAGE: line 9: Message #10
CONSOLE MESSAGE: line 9: Message #11
CONSOLE MESSAGE: line 9: Message #12
CONSOLE MESSAGE: line 9: Message #13
CONSOLE MESSAGE: line 9: Message #14
CONSOLE MESSAGE: line 9: Message #15
CONSOLE MESSAGE: line 9: Message #16
CONSOLE MESSAGE: line 9: Message #17
CONSOLE MESSAGE: line 9: Message #18
CONSOLE MESSAGE: line 9: Message #19
CONSOLE MESSAGE: line 9: Message #20
CONSOLE MESSAGE: line 9: Message #21
CONSOLE MESSAGE: line 9: Message #22
CONSOLE MESSAGE: line 9: Message #23
CONSOLE MESSAGE: line 9: Message #24
CONSOLE MESSAGE: line 9: Message #25
CONSOLE MESSAGE: line 9: Message #26
CONSOLE MESSAGE: line 9: Message #27
CONSOLE MESSAGE: line 15: Repeating message
CONSOLE MESSAGE: line 15: Repeating message
CONSOLE MESSAGE: line 15: Repeating message
CONSOLE MESSAGE: line 15: Repeating message
CONSOLE MESSAGE: line 15: Repeating message
CONSOLE MESSAGE: line 15: Repeating message
CONSOLE MESSAGE: line 15: Repeating message
CONSOLE MESSAGE: line 15: Repeating message
CONSOLE MESSAGE: line 15: Repeating message
CONSOLE MESSAGE: line 15: Repeating message
CONSOLE MESSAGE: line 15: Repeating message
CONSOLE MESSAGE: line 15: Repeating message
CONSOLE MESSAGE: line 15: Repeating message
CONSOLE MESSAGE: line 9: Message #0
CONSOLE MESSAGE: line 9: Message #1
CONSOLE MESSAGE: line 9: Message #2
Verifies viewport correctly shows and hides messages while logging and scrolling.

Max visible messages count: 13, active count: 25

Running: addSmallCount
Messages shown: 3, hidden: 0

Running: addMaxVisibleCount
Messages shown: 13, hidden: 0

Running: addMaxActiveCount
Is at bottom: true, should stick: true
Messages shown: 25, hidden: 0

Running: addMoreThanMaxActiveCount
Is at bottom: true, should stick: true
Messages shown: 28, hidden: 3

Running: scrollUpWithinActiveWindow
Is at bottom: true, should stick: true
Messages shown: 0, hidden: 0

Running: scrollUpToPositionOutsideOfActiveWindow
Is at bottom: true, should stick: true
Messages shown: 3, hidden: 3

Running: logRepeatingMessages
Messages shown: 1, hidden: 0

Running: reorderingMessages
Logging 3 messages
Messages shown: 3, hidden: 0
Swapping messages 0 and 1
Messages shown: 0, hidden: 0
